move decorators; fix type issue
